apps/api/pyproject.toml
[project]
name = "api"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sentry-sdk[fastapi]>=2.18",
    "PyGithub>=2.4",
    "slowapi>=0.1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/api"]
